+++ packages/quality-config/CHANGELOG.md
@@ -1,6 +1,6 @@
 # @busirocket/quality-config
 
-## Unreleased
+## 0.5.0
 
 ### Minor Changes